+++ plugins/main/common/constants.ts
@@ -326,7 +326,7 @@ export const WAZUH_SAMPLE_DATA_CATEGORIES_TYPE_DATA = {
   })),
   [WAZUH_SAMPLE_ALERTS_CATEGORY_AUDITING_POLICY_MONITORING]: [
     {
-      rootcheck: true,
+      'wazuh-rootcheck': true,
     },
     {
       audit: true,
@@ -518,7 +518,7 @@ export enum WAZUH_MODULES_ID {
   GOOGLE_CLOUD_PLATFORM = 'gcp',
   POLICY_MONITORING = 'pm',
   SECURITY_CONFIGURATION_ASSESSMENT = 'sca',
-  AUDITING = 'audit',
+  AUDITING = 'azure-audit',
   VULNERABILITIES = 'vuls',
   DOCKER = 'docker',
   MITRE_ATTACK = 'mitre',