+++ plugins/main/public/components/common/data-source/pattern/events/fim/fim-data-source.ts
@@ -2,8 +2,8 @@ import { tFilter } from '../../../index';
 import { DATA_SOURCE_FILTER_CONTROLLED_FIM_RULE_GROUP } from '../../../../../../../common/constants';
 import { EventsDataSource } from '../events-data-source';
 
-const FIM_GROUP_KEY = 'rule.groups';
-const FIM_GROUP_VALUE = 'syscheck';
+const FIM_GROUP_KEY = 'wazuh.integration.decoders';
+const FIM_GROUP_VALUE = 'wazuh-fim';
 
 export class FIMDataSource extends EventsDataSource {
   constructor(id: string, title: string) {