+++ plugins/main/public/components/common/data-source/pattern/events/office-365/office-365-data-source.ts
@@ -2,7 +2,7 @@ import { tFilter } from '../../../index';
 import { DATA_SOURCE_FILTER_CONTROLLED_OFFICE_365_RULE_GROUP } from '../../../../../../../common/constants';
 import { EventsDataSource } from '../events-data-source';
 
-const OFFICE_365_GROUP_KEY = 'rule.groups';
+const OFFICE_365_GROUP_KEY = 'wazuh.integration.decoders';
 const OFFICE_365_GROUP_VALUE = 'office365';
 
 export class Office365DataSource extends EventsDataSource {